requirements.txt
anthropic>=0.40.0
langgraph>=0.2.50
langchain-core>=0.3.0
langfuse>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
apscheduler>=3.10.4
jinja2>=3.1.4
python-dotenv>=1.0.0
python-dateutil>=2.9.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
